routes split by schemas, fixing and adding all details for authentication algorythms
