+++ index.js
@@ -17,13 +17,19 @@ app.use(bodyParser.json())
 app.use(bodyParser.urlencoded({extended: true}))
 app.use(passport.initialize())
 app.use(expressValidator())
-const routes = require('./routes')
+const userRoutes = require('./routes/userRoutes')
+const listRoutes = require('./routes/listRoutes')
+const officeRoutes = require('./routes/officeRoutes')
+const orderRoutes = require('./routes/orderRoutes')
 const loginRoute = require('./middlewares/loginRoutes')
 
 app.use(errorHandlers.productionErrors)
 
 app.use('/', loginRoute)
-app.use('/', routes)
+app.use('/', userRoutes)
+app.use('/', officeRoutes)
+app.use('/', listRoutes)
+app.use('/', orderRoutes)
 
 app.use(errorHandlers.notFound)
 app.use(errorHandlers.catchErrors)
